dataset.py
import os
import torch
import torch.utils.data as data
import numpy as np
from PIL import Image

class MLGCNDataset(data.Dataset):
    def __init__(self, data_path, transform=None, target_transform=None):
        self.data_path = data_path
        self.transform = transform
        self.target_transform = target_transform
        
        # Load image paths and labels
        self.imgs = []
        self.labels = []
        
        # Example implementation - override in subclasses
        # Load your actual data here
        self._load_data()
    
    def _load_data(self):
        raise NotImplementedError("Subclasses must implement _load_data")
    
    def __getitem__(self, index):
        img_path = self.imgs[index]
        target = self.labels[index]
        
        img = Image.open(img_path).convert('RGB')
        if self.transform is not None:
            img = self.transform(img)
        
        if self.target_transform is not None:
            target = self.target_transform(target)
        
        return img, target
    
    def __len__(self):
        return len(self.imgs)

class TrainDataset(MLGCNDataset):
    def _load_data(self):
        # Example implementation for COCO or similar datasets
        img_dir = os.path.join(self.data_path, 'images', 'train')
        label_file = os.path.join(self.data_path, 'labels', 'train_labels.npy')
        
        if os.path.exists(label_file):
            self.labels = np.load(label_file)
        else:
            # Placeholder for demo
            self.labels = np.zeros((10, 80), dtype=np.float32)
        
        # Get image paths
        if os.path.exists(img_dir):
            self.imgs = [os.path.join(img_dir, img_name) for img_name in os.listdir(img_dir)
                        if img_name.endswith(('.jpg', '.jpeg', '.png'))][:len(self.labels)]
        else:
            # Placeholder for demo
            self.imgs = [f"placeholder_{i}.jpg" for i in range(len(self.labels))]
            print(f"Warning: Image directory {img_dir} not found. Using placeholders.")

class ValDataset(MLGCNDataset):
    def _load_data(self):
        # Example implementation for COCO or similar datasets
        img_dir = os.path.join(self.data_path, 'images', 'val')
        label_file = os.path.join(self.data_path, 'labels', 'val_labels.npy')
        
        if os.path.exists(label_file):
            self.labels = np.load(label_file)
        else:
            # Placeholder for demo
            self.labels = np.zeros((10, 80), dtype=np.float32)
        
        # Get image paths
        if os.path.exists(img_dir):
            self.imgs = [os.path.join(img_dir, img_name) for img_name in os.listdir(img_dir)
                        if img_name.endswith(('.jpg', '.jpeg', '.png'))][:len(self.labels)]
        else:
            # Placeholder for demo
            self.imgs = [f"placeholder_{i}.jpg" for i in range(len(self.labels))]
            print(f"Warning: Image directory {img_dir} not found. Using placeholders.")